==> 02-model.Rmd <==
---
title: "COVID-19 - Curva Epidemiológica"
author: "Emanuel Diego S Penha"
output:
  pdf_document:
    latex_engine: xelatex
    number_sections: yes
    toc: yes
date: '`r format(Sys.Date(), "%B %d, %Y")`'
header-includes:
- \usepackage[brazil]{babel}
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE,
                      fig.width = 10,
                      echo = FALSE)

options(scipen = 10, digits = 6)
```

# Informações

- Tentativa de implementar algo como visto em 
https://rviews.rstudio.com/2020/03/05/covid-19-epidemiology-with-r/

# Carregar pacotes

```{r}
library("tidyverse")
library("lubridate")
library("earlyR")
library("EpiEstim")
theme_set(theme_minimal())

covid <- read_csv2("dados/COVID19.csv", local = locale(encoding = "latin1")) %>%
#	rename(estado = sigla) %>%
	rename(dia = data) %>%
#	mutate(dia = dmy(dia)) %>%
	filter(dia >= ymd("2020-02-26"))
#	mutate(casosAcumulados = cases) %>%
#	mutate(obitosAcumulados = deaths) %>%

bra <- covid %>%
  group_by(dia) %>%
  summarise(confirmados = sum(casosAcumulados, na.rm = TRUE),
            óbitos = sum(obitosAcumulados, na.rm = TRUE)) 

bra
# bra <- mutate(bra, casos = if_else(casos == 0, as.numeric(NA), casos)) 
```

# Variáveis que serão usadas depois
```{r}
hoje <- Sys.Date()

ultimo_dia <- max(covid$dia)

primeiro_dia <- min(covid$dia)

```

# Modelo exponencial simples

```{r}
# modelando
me <- with(filter(bra, dia > ultimo_dia - 15), lm(log(confirmados) ~ dia))
summary(me)

periodo_pred <- seq(primeiro_dia, ultimo_dia + 7, 1)
predicao <- exp(predict(me,list(dia=periodo_pred)))
pre <- tibble(dia = periodo_pred, pred = predicao)

bra_pred <- left_join(pre, bra, by = "dia")

tail(bra_pred, 15)


# gráfico em barras
ggplot(bra_pred, aes(x = dia, y = confirmados)) +
	geom_col() 

# mostrando curva exponencial 
filter(bra_pred)  %>% na.omit %>%
ggplot(aes(x = dia, y = confirmados)) +
	geom_col() +
	xlim(primeiro_dia, ultimo_dia) +
	geom_line(aes(y = pred))

# mostrando curva exponencial e predição
bra_pred %>%
ggplot(aes(x = dia, y = confirmados)) +
	geom_col() +
	geom_line(aes(y = pred))

ggplot(bra_pred, aes(x = dia, y = confirmados)) +
  geom_point() +
  geom_line(aes(y = pred)) + 
  labs(title = 'Predição do número casos para os próximos 7 dias - COVID-19',
       subtitle = paste0('último dia: ', ultimo_dia, ', predição até: ', ultimo_dia + 7) ) 

ggplot(bra_pred, aes(x = dia, y = confirmados)) +
  scale_y_continuous(trans='log10') + 
  geom_point() +
  geom_line(aes(y = pred)) + 
  labs(title = 'Predição do número casos para os próximos 7 dias - COVID-19',
       subtitle = paste0('último dia: ', ultimo_dia, ', predição até: ', ultimo_dia + 7) ) 
  

```

# Usando pacotes específicos para Epidemiologia


## Estimando incidência
- Para calcular a incidência subtraí o total de casos de um dia pelo o do dia anterior
- Aqui também criei um vetor com as datas de registro de cada caso

```{r}
library('incidence')
library('lubridate')

incid <- vector()
datas <- vector()

incid <- 1

for(i in 2:nrow(bra)){
	print(i)
	j <- bra$confirmados[i] - bra$confirmados[i-1]
 incid <- c(incid, j)
}

for(i in 1:nrow(bra)){
 datas <- c(datas, rep(bra$dia[i], incid[i]))
}

datas <- as_date(datas)

bra <- mutate(bra, incidencia = incid)
bra_pred <- left_join(bra_pred, bra)

```

Alguns gráficos de incidência


```{r}

ggplot(bra, aes(dia, incidencia)) + 
	geom_col() 

ggsave('plots/incidência_barra.png', device = 'png')

ggplot(bra, aes(dia, incidencia)) + 
	geom_line() 

ggsave('plots/incidência_linha.png', device = 'png')
```

Criação do objeto do tipo incidência e cálculo de Infectividade Lambda

```{r}
#hj <- as.Date("2020-03-30")
#incid_p <- incidence(dates = datas, last_date = hj)
incid_p <- incidence(dates = datas)

# os valores de média e desvio padrão são dados no post, 
# são derivados 
# o seq(0, 20) foi retirado do manual do pacote 

discrete_si_distr <- discr_si(seq(0, 20), 5, 3.4)

lambda <- overall_infectivity(incid_p, discrete_si_distr)

bra_pred <- bra_pred %>% filter(dia <= ultimo_dia)

bra_pred <- mutate(bra_pred, lambda = lambda) 

ggplot(bra_pred, aes(dia, incidencia)) +
	geom_col(fill = "steelblue") +
	geom_line(aes(y = lambda),  size = 2, color = 'red') +
	labs(title = "Incidência e Força de Infecção (Lambda) para COVID-19, Brasil")

ggsave('plots/incidência_lambda.png', device = 'png')

```

R

```{r}
# pacote earlyR
#res <- earlyR::get_R(incid_p, si_mean = 5, si_sd = 3.4)
#res
#plot(res)
```
